% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingResults.R
\name{addTrypticInfo}
\alias{addTrypticInfo}
\title{Function for duplicating peptides if they occur multiple times in the
same protein}
\usage{
addTrypticInfo(pepDf, x, sign)
}
\arguments{
\item{pepDf}{Data.frame for plotting protein in woods plot as created within
\code{separateProteins} or \code{separatePeptides}}

\item{x}{Data.frame for plotting protein in woods plot not yet altered by
\code{separateProteins} or \code{separatePeptides}}

\item{sign}{A character variable to separate the 'isTryptic' column in
\code{x}}
}
\value{
\code{plotData} data.frame
}
\description{
Function splits up \code{isTryptic} information in case
quantities mapping to multiple positions
}
