% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModels.R
\name{analyzeLiPPepData}
\alias{analyzeLiPPepData}
\title{Fitting model for accessibility changes in LiP peptides.}
\usage{
analyzeLiPPepData(spectroList, annotS, infoCondition="Condition",
formulaRUV="Y~XPep+XProt", formulaContrast=NULL, lowRUV=c(-1e9, 0, 0),
upRUV=c(Inf, Inf, Inf), addRUVbounds=FALSE, LiPonly=FALSE, withHT=FALSE)
}
\arguments{
\item{spectroList}{A list of matrices, containing peptide/protein quantities.
Rows represent features and columns refer to the samples. Names of the list
items should be set to "LiPPep", "TrpPep" and "TrpProt" (if the LiP only
version is run, names should refer to "LiPPep" and "LiPProt").}

\item{annotS}{A data.frame containing sample annotation. Rows are samples and
must match to columns of \code{spectroList}.}

\item{infoCondition}{A character string providing column name of
\code{annotS} in which condition is provided. Default is 'Condition'.}

\item{formulaRUV}{A character string or formula defining the RUV models.
Default is defined as 'Y~XPep+XProt'.}

\item{formulaContrast}{A character string or formula defining the contrast
models. If 'NULL' will be set to 'Y~\code{infoCondition}.}

\item{lowRUV}{A numeric vector defining lower boundaries of the coefficients
of the RUV models. Elements refer to definition of \code{formulaRUV}.
Default is defined as 'c(-Inf, 0, 0)'.}

\item{upRUV}{A numeric vector defining upper boundaries of the coefficients
of the RUV models. Elements refer to definition of \code{formulaRUV}.
Default is defined as 'c(Inf, Inf, Inf)'.}

\item{addRUVbounds}{A boolean value, if set to 'TRUE' as many bounds as
additionally needed in each RUV model are added to \code{lowRUV} and
\code{upRUV}. Added boundaries are automatically set to -Inf for
\code{lowRUV} and Inf for \code{upRUV}. Important to set to 'TRUE', if you
are for example also running batch correction in the RUV model.}

\item{LiPonly}{A boolean value to set to 'TRUE' if you are running the
LiPonly version of the package and not providing trypsin-only data. If set
to TRUE' RUV boundaries will be adjusted, \code{lowRUV} will be set to
c(-1e9, 0) and \code{RUV} will be set to c(Inf, Inf). Additionally,
\code{formulaRUV} will be adjusted. In case you want to add further
variables to the models, please use the \code{runModel} function. Default is
'FALSE'.}

\item{withHT}{A boolean value to set to 'TRUE' if LiPPep should only be
corrected for TrpProt only. If set to TRUE' RUV boundaries will be adjusted,
\code{lowRUV} will be set to c(-1e9, 0) and \code{RUV} will be set to
c(Inf, Inf). Additionally, \code{formulaRUV} will be adjusted. In case you
want to add further variables to the models, please use the \code{runModel}
function. Default is 'FALSE'.}
}
\description{
Function to build linear regression models for fitting MS data
and retrieving structural variation between different conditions.
}
