% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModels.R
\name{AnalyzeTrpPepData}
\alias{AnalyzeTrpPepData}
\title{Fitting model for PK-indendepent changes in Trp peptides}
\usage{
AnalyzeTrpPepData(spectroList, annotS, infoCondition = "Condition",
formulaBVLS = "Y~XProt", formulaOLS = NULL,
lowBVLS = c(-1e9, 0), upBVLS = c(Inf, Inf), addBVLSbounds = F)
}
\arguments{
\item{spectroList}{a list of matrices, containing peptide/protein quantities.
Rows represent features and columns refer to the samples.
Names of the list items should be set to "LiPPep", "TrpPep" and "TrpProt"
(if the LiP only version is run, names should refer to "LiPPep" and "LiPProt")}

\item{annotS}{a data.frame object containing sample annotation.
Rows are samples and must match to columns of \code{spectroList}.}

\item{infoCondition}{a character, providing column name of \code{annotS} in which condition is provided.
Default is 'Condition'.}

\item{formulaBVLS}{a character string or formula defining the BVLS models.
Default is defined as 'Y~XProt'.}

\item{formulaOLS}{a character string or formula defining the OLS models.
If 'NULL' will be set to 'Y~\code{infoCondition}.}

\item{lowBVLS}{a numeric vector defining lower boundaries of the coefficients of the BVLS models.
Elements refer to definition of \code{formulaBVLS}.
Default is defined as 'c(-Inf, 0)'.}

\item{upBVLS}{a numeric vector defining upper boundaries of the coefficients of the BVLS models.
Elements refer to definition of \code{formulaBVLS}.
Default is defined as 'c(Inf, Inf)'.}

\item{addBVLSbounds}{a boolean value, if TRUE as many bounds as additionally needed
in each BVLS model are added to \code{lowBVLS} and \code{upBVLS}.
Added boundaries are automatically set to -Inf for \code{lowBVLS} and Inf for \code{upBVLS}.}
}
\description{
Function to build linear regression models for fitting MS data

add later
}
