% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModels.R
\name{AnalyzeTrpProtData}
\alias{AnalyzeTrpProtData}
\title{Fitting model for protein abundance changes}
\usage{
AnalyzeTrpProtData(spectroList, annotS, annotPP,
infoCondition="Condition", infoProtName="Protein", formulaBVLS=NULL,
formulaOLS=NULL, lowBVLS=NULL, upBVLS=NULL, addBVLSbounds=FALSE,
LiPonly=FALSE)
}
\arguments{
\item{spectroList}{A list of matrices, containing peptide/protein quantities.
Rows represent features and columns refer to the samples. Names of the list
items should be set to "LiPPep", "TrpPep" and "TrpProt" (if the LiP only
version is run, names should refer to "LiPPep" and "LiPProt").}

\item{annotS}{A data.frame containing sample annotation. Rows are samples and
must match to columns of \code{spectroList}.}

\item{annotPP}{A data.frame with peptide and protein annotatioon. Rows are
features and must match to the row.names of SpectroList/QuantityMatrix.}

\item{infoCondition}{A character string providing column name of
\code{annotS} in which condition is provided. Default is 'Condition'.}

\item{infoProtName}{A character string providing column name of
\code{annotPP} in which the protein names are provided. Default is 'Protein'.}

\item{formulaBVLS}{A character string or formula defining the BVLS models.
Default is defined as 'NULL'.}

\item{formulaOLS}{A character string or formula defining the OLS models.
If 'NULL' will be set to 'Y~\code{infoCondition}}

\item{lowBVLS}{A numeric vector defining lower boundaries of the coefficients
of the BVLS models. Elements refer to definition of \code{formulaBVLS}.
Default is defined as 'NULL'.}

\item{upBVLS}{A numeric vector defining upper boundaries of the coefficients
of the BVLS models. Elements refer to definition of \code{formulaBVLS}.
Default is defined as 'NULL'.}

\item{addBVLSbounds}{A boolean value, if set to 'TRUE' as many bounds as
additionally needed in each BVLS model are added to \code{lowBVLS} and
\code{upBVLS}. Added boundaries are automatically set to -Inf for
\code{lowBVLS} and Inf for \code{upBVLS}. Important to set to 'TRUE', if you
are for example also running batch correction in the BVLS model.}

\item{LiPonly}{A boolean value to set to 'TRUE' if you are running the
LiPonly version of the package and not providing trypsin-only data.}
}
\description{
Function to build linear regression models for fitting MS data
and retrieving protein abundance variation between different conditions.
}
