% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModels.R
\name{calcualtePvalAfterRUV}
\alias{calcualtePvalAfterRUV}
\title{Estimate p-values from contrast model taking degrees of freedom used
in RUV step into account}
\usage{
calcualtePvalAfterRUV(LM, coeffTval, dfRUV)
}
\arguments{
\item{LM}{A single linear regression model originating from contrast step}

\item{coeffTval}{A character variable giving the column name were t-values
are provided if \code{summary.lm(mRUV)}.}

\item{dfRUV}{A numberic value providing the number of degrees of freedom used
in the RUV models.}
}
\value{
Returns a data.frame with p-values.
}
\description{
Estimate p-values from contrast model taking degrees of freedom used
in RUV step into account
}
