% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingAndSorting.R
\name{convert2Matrix}
\alias{convert2Matrix}
\title{Extracting a single peptide/protein matrix from Spectronaut report}
\usage{
convert2Matrix(spectroOut, quantValue, rows, cols)
}
\arguments{
\item{spectroOut}{Spectronaut report exported using the Spectronaut schema
'SpectroSchema_LiPAnalyzerOut'.}

\item{quantValue}{A character string or numeric giving the column name or
column number from the Spectronaut report in which quantities that should be
writen into matrix are located.If there are different quantities for the
same \code{rows} the mean of these quantValue is used.}

\item{rows}{A character string or numeric giving the column name or
column number from the  Spectronaut report where the names of the quantities
writen into matrix are located.}

\item{cols}{A character string or numeric giving the column name or
column number from the  Spectronaut report where the sample or file names are
located.}
}
\value{
Returns a matrix with quantities, \code{rows} represent features
and \code{cols} refer to the samples.
}
\description{
This function exports quantities from the Spectronaut report
and writes them into a matrix. Rows are peptides/proteins and columns are
samples.
}
