% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingResults.R
\name{exportWoodsPlots}
\alias{exportWoodsPlots}
\title{Function for exporting woods plots and writing them into a .pdf file}
\usage{
exportWoodsPlots(plotList, lList, file)
}
\arguments{
\item{plotList}{A list of woods plots created with the function
\code{plottingProtein} were the variable \code{export} was set to 'TRUE.}

\item{lList}{A numeric variable providing the length of \code{plotList}.}

\item{file}{A character string providing export path and file name of pdf
with plots.}
}
\value{
'NULL'
}
\description{
Function for exporting woods plots and writing them into a .pdf file
}
