% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModels.R
\name{extractContrast}
\alias{extractContrast}
\title{Extracting information from contrast models}
\usage{
extractContrast(mContrast, formulaContrast, dfRUV, coeffPval="Pr(>|t|)",
coeffTval="t value")
}
\arguments{
\item{mContrast}{A list of contrast models estimated using ordinary least square}

\item{formulaContrast}{A formula used to create \code{modelMat} for
contrast models.}

\item{dfRUV}{A numberic value providing the number of degrees of freedom used
in the RUV models.}

\item{coeffPval}{A character variable giving the column name were p-values
are provided if \code{summary.lm(mRUV)}.
Default is "Pr(>|t|)".}

\item{coeffTval}{A character variable giving the column name were t-values
are provided if \code{summary.lm(mRUV)}.
Default is "t value".}
}
\value{
A list were the first element is a data.frame with residuals from
from the contrast models.
}
\description{
Function to extract the coefficients and residuals from each
RUV model. If RUV was run before the contrast model, the p-values estimation
taks into account the degrees of freedom already used by the RUV step.
}
