% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingAndSorting.R
\name{extractData}
\alias{extractData}
\title{Creating peptide and protein matrices from a XXX report.}
\usage{
extractData(libraryLiP, libraryTrp=NULL, sampleName, quantName,
quantValue, protValue, LiPonly=FALSE)
}
\arguments{
\item{libraryLiP}{Exported report of LiP data. Has to include columns with
the name of the samples (\code{sampleName}), name of quantities which will be
the row names of the quantity matrices (e.g. peptide names,
\code{quantName}), quantities matching the \code{quantName} (e.g. peptide
quantities, \code{quantValue}) and the protein quantities (\code{protValue}).}

\item{libraryTrp}{Exported report of the Trp data. Has to include the same
columns as the \code{libraryLiP}. Does not have to be included if the LiPonly
mode is run (\code{LiPonly} = TRUE).}

\item{sampleName}{A character string or numeric giving the column name or
column number in which sample names are defined in the Spectronaut report.}

\item{quantName}{A character string or numeric giving the column name or
column number in which the row names of the quantity matrices created in this
function are defined.}

\item{quantValue}{A character string or numeric giving the column name or
column number with quantities to infer structural alterations from (e.g.
peptide quantities), should match the \code{quantName}.}

\item{protValue}{A character string or numeric giving the column name or
column number from which protein quantities should be taken from.}

\item{LiPonly}{A boolean value, set to 'TRUE' if you want to run the LiPonly
version o the package and not providing trypsin-only data.Default is set to
'FALSE'.}
}
\value{
Per default a list of three matrices is returned, containing the LiP
peptide/precursor, Trp peptide/precursor and Trp protein quantities. If run
in \code{LiPonly} mode will return a list of two matrcies with the LiP
peptide/precursor and Trp protein quantities. Rows represent features and
columns refer to the samples.
}
\description{
This function writes peptide and protein quantities from a
report into quantity matrices.
}
