% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingAndSorting.R
\name{extractDataFromSpectro}
\alias{extractDataFromSpectro}
\title{Creating peptide and protein matrices from the Spectronaut report.}
\usage{
extractDataFromSpectro(spectroLiP, spectroTrp=NULL,
analysisLvl="Peptide", sampleName="R.FileName", quantName=NULL,
quantValue=NULL, protValue="PG.Quantity", LiPonly=FALSE)
}
\arguments{
\item{spectroLiP}{Spectronaut report of LiP data. Spectronaut report should
have been exported using the Spectronaut schema
'SpectroSchema_LiPAnalyzerOut'.}

\item{spectroTrp}{Spectronaut report of Trp data. Spectronaut report should
have been exported using the Spectronaut schema
'SpectroSchema_LiPAnalyzerOut'.}

\item{analysisLvl}{A character string defining the level on which the
peptide/protein quantities should be exported. Is set to 'Peptide' by
default, can alternatively be set to 'ModifiedPeptide' or 'Precursor'.This
will define the column which will be converted into row.names of the
matrices ('Peptide'='PEP.Quantity', ModifiedPeptide'='FG.Quantity',
'Precursor'='FG.Quantity'). If you are using an alternative input format
and/or want to choose a different column, please set this variable to
'individual' and specify the column in the variable \code{quantName}. You then
also have to define the \code{'quantValue'}.}

\item{sampleName}{A character string or numeric giving the column name or
column number in which sample names are defined in the Spectronaut report.
Set to R.FileName' by default.}

\item{quantName}{A character string or numeric giving the column name or
column number in which the row names of the quantity matrices are defined.
This parameter will only be evaluated if \code{analysisLvl} is set to
individual'.}

\item{quantValue}{A character string or numeric giving the column name or
column number from which peptide/precursor quantities should be taken from.
Defined as 'NULL' by default, resulting in the function automatically setting
\code{quantValue} based on the \code{AnalysisLvl} ('Peptide'='PEP.Quantity',
'ModifiedPeptide'='FG.Quantity', 'Precursor'='FG.Quantity').}

\item{protValue}{A character string or numeric giving the column name or
column number from which protein quantities should be taken from. Set to
'PG.Quantity' by default.}

\item{LiPonly}{A boolean value, set to 'TRUE' if you want to run the LiPonly
version o the package and not providing trypsin-only data.Default is set to
'FALSE'.}
}
\value{
Per default a list of three matrices is returned, containing the LiP
peptide/precursor, Trp peptide/precursor and Trp protein quantities. If run
in \code{LiPonly} mode will return a list of two matrcies with the LiP
peptide/precursor and Trp protein quantities. Rows represent features and
columns refer to the samples.
}
\description{
This function writes peptide and protein quantities from a
Spectronaut report into a matrix.Data is exported from both the LiP and Trp
Spectronaut reports per default. Alternatively, function can be run in
\code{LiPonly} mode, only requiring LiP data.
}
