% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingAndSorting.R
\name{extractMSData}
\alias{extractMSData}
\title{Creating quantity matrices from a MS report.}
\usage{
extractMSData(reportLiP, reportTrP=NULL, sampleName, quantName,
quantValue, protValue, LiPonly=FALSE)
}
\arguments{
\item{reportLiP}{MS report of LiP data as data.frame or matrix. Has to
include columns with the name of the samples (\code{sampleName}), name of
quantities which will be the row names of the quantity matrices (e.g. peptide
names, \code{quantName}), quantities matching the \code{quantName} (e.g.
peptide quantities, \code{quantValue}) and the protein quantities
(\code{protValue}).}

\item{reportTrP}{MS report of the TrP data as data.frame or matrix. Has to
include the same columns as the \code{reportLiP}. Does not have to be
provided if the 'LiPonly' mode is run. For this set (\code{LiPonly} = TRUE).
If LiP and TrP data were processed together, please separate them into two
input files.}

\item{sampleName}{A character string or numeric giving the column name or
column number in which sample names are provided in the MS report.}

\item{quantName}{A character string or numeric giving the column name or
column number in which the row names of the quantity matrices created in this
function are provided (e.g. peptide, modified peptide, precursor).}

\item{quantValue}{A character string or numeric giving the column name or
column number with quantities to infer structural alterations from (e.g.
peptide quantities, modified peptide quantities, precursors). These have to
match to the ID provided in \code{quantName}.}

\item{protValue}{A character string or numeric giving the column name or
column number were protein quantities are provided. If a peptides matches to
the same protein several times, the protein name should be provided each
time, separated by ','. If a peptide maps to the multiple proteins, these
different proteins can be provided by separating them by ';'.}

\item{LiPonly}{A boolean value, set to 'TRUE' if you want to run the LiPonly
version o the package and not use trypsin-only data. Default is set to
'FALSE'.}
}
\value{
If run in default mode, a list of three matrices is returned:
\itemize{
  \item 'LiPPep': LiP peptide quantities (or alternatively the modified
  peptide/ precursor/other quantities, dependent on \code{quantName} and
  \code{quantValue})
  \item 'TrPPep': TrP peptide quantities (or alternatively the modified
  peptide/precursor/other quantities, dependent on \code{quantName} and
  \code{quantValue})
  \item 'TrpProt': TrP protein quantities
  }

If run in \code{LiPonly} mode, a list of two matrices is returned:
\itemize{
  \item 'LiPPep': LiP peptide quantities (or alternatively the modified
  peptide/precursor/other quantities, dependent on \code{quantName} and
  \code{quantValue})
  \item 'LiPProt': LiP protein quantities
}

Rows represent features and columns samples.
}
\description{
Extract the quantity of interest, e.g. peptide intensities, and
protein quantities from a MS report/ MS reports.
}
