% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModels.R
\name{extractRUV}
\alias{extractRUV}
\title{Extracting information from RUV models}
\usage{
extractRUV(mRUV, samples)
}
\arguments{
\item{mRUV}{A list of RUV models estimated using bounded variable least
square regression}

\item{samples}{A character vector providing sample names for the model.}
}
\value{
A list were the first element is a data.frame with residuals from
from the RUV models.
}
\description{
Function to extract the coefficients and residuals from each
RUV model
}
