% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingAndSorting.R
\name{extractSpectroData}
\alias{extractSpectroData}
\title{Creating quantity matrices from a Spectronaut report.}
\usage{
extractSpectroData(spectroLiP, spectroTrp=NULL, analysisLvl="Peptide",
sampleName="R.FileName", LiPonly=FALSE)
}
\arguments{
\item{spectroLiP}{Spectronaut report of LiP data. Spectronaut report has to
be exported using the Spectronaut schema SpectroSchema_LiPAnalyzerOut'.}

\item{spectroTrp}{Spectronaut report of Trp data. Spectronaut report has to
be exported using the Spectronaut schema 'SpectroSchema_LiPAnalyzerOut'.
If LiP and Trp data were processed together in Spectronaut, please separate
them into two input files.}

\item{analysisLvl}{A character string defining the level on which the
peptide/protein quantities should be exported. Is set to 'Peptide' by
default, can alternatively be set to 'ModifiedPeptide' or 'Precursor'.This
will define the column which will be converted into row.names of the
matrices ('Peptide'='PEP.Quantity', ModifiedPeptide'='FG.Quantity',
'Precursor'='FG.Quantity'). If you are want to choose a different column,
 else, please use the function /code{extractMSData}.}

\item{sampleName}{A character string or numeric giving the column name or
column number in which sample names are provided in the Spectronaut report.
Set to R.FileName' by default.}

\item{LiPonly}{A boolean value, set to 'TRUE' if you want to run the LiPonly
version o the package and not use trypsin-only data. Default is set to
'FALSE'.}
}
\value{
If run in default mode, a list of three matrices is returned:
\itemize{
  \item 'LiPPep': LiP peptide quantities (or alternatively the modified
  peptide/ precursor/other quantities, dependent on \code{quantName} and
  \code{quantValue})
  \item 'TrpPep': Trp peptide quantities (or alternatively the modified
  peptide/precursor/other quantities, dependent on \code{quantName} and
  \code{quantValue})
  \item Trp protein quantities
  }

If run in \code{LiPonly} mode, a list of two matrices is returned:
\itemize{
  \item 'LiPPep': LiP peptide quantities (or alternatively the modified
  peptide/precursor/other quantities, dependent on \code{quantName} and
  \code{quantValue})
  \item 'LiPProt': LiP protein quantities
}

Rows represent features and columns samples.
}
\description{
Extract the quantity of interest, e.g. peptide intensities,
and protein quantities from a Spectronaut report/Spectronaut reports.
Data must have been exported using the Spectronaut schema
'SpectroSchema_LiPAnalyzerOut', else, please use the function
/code{extractMSData}.
}
