% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterSpectroOut.R
\name{filterForHTonly}
\alias{filterForHTonly}
\title{Filtering data for running 'HTonly' mode}
\usage{
filterForHTonly(
  quantityList,
  annotPP,
  annotS,
  filterNA,
  maxNAs,
  infoCondition,
  nameIDQuant,
  nameProtQuant,
  infoTryptic,
  nameFT
)
}
\arguments{
\item{quantityList}{A list of matrices, containing quantities of interest
(e.g. peptide, modified peptide, precursor) and protein abundances. Rows
represent features and columns samples. Sample names between the different
matrices contained in the list should match. Not matching samples will be
removed.
Output from \code{extractMSData} or \code{extractSpectroData} is in
the correct format (except of potentially not matching sample names, then
these need to be adjusted manually.)
\code{quantityList} must include the following matrices
\itemize{
  \item 'LiPPep': LiP peptide quantities (or alternatively the modified
  peptide/ precursor/other quantities, dependent on \code{quantName} and
  \code{quantValue})
  \item 'TrPPep': TrP peptide quantities (or alternatively the modified
  peptide/precursor/other quantities, dependent on \code{quantName} and
  \code{quantValue})
  \item 'TrPProt': TrP protein quantities
  }}

\item{annotPP}{A data.frame with peptides (/modified peptides/precursors) and
protein annotation. Rows are features and must match to the row.names of
quantityList. Must include all columns required for filtering
data according to settings provided to this function. Per default, the output
from \code{getPepProtAnnot} can be given here.}

\item{annotS}{A data.frame object containing sample annotation. Rows are
samples and must match to columns of \code{quantityList}.
If \code{filterNA='byCondition'} it needs to contain column providing the
conditions/groups.}

\item{filterNA}{A character variable defining type of NA filtering applied.
Can be set to c('all', 'byCondition' or 'none').
\itemize{
  \item 'all': Filtering of NAs will be applied on all samples, independent
  of their condition. \code{maxNA} defines the number of NAs allowed per
  quantity over all samples.
  \item 'byCondition': Filtering of NAs will be applied on samples based on
  their condition.  \code{maxNA} defines the number of NAs allowed per
  quantity over all samples of each individual condition. It is advised to
  only use 'byCondition', if the condition you are providing is a
  categorical variable and not continuous.
  \item 'none': No NA filtering will be applied. This my cause issues if
  running RUV and/or contrast models subsequently.
  }}

\item{maxNAs}{A numeric value, defining maximal number of NAs allowed.
Default is '0'.}

\item{infoCondition}{A character string or numeric giving the column name
or column number of \code{annotS} in which condition is provided. Required,
if \code{filterNA = 'byCondition'}.
Default is 'Condition'.}

\item{nameIDQuant}{A character string giving column of \code{annotPP} which
matches to the row.names of \code{quantityList} or \code{quantityMatrix}.
Default is 'quantID'.}

\item{nameProtQuant}{A character string giving column of \code{annotPP} which
protein names are provided. If a peptides matches to the same protein several
times, the protein name should be provided each time, separated by ','. If a
peptide maps to the multiple proteins, these different proteins can be
provided by separating them with ';'.
Default is 'Protein'.}

\item{infoTryptic}{A character string or numeric giving the column name or
column number of \code{annotPP} in which digest type is provided.
Default is 'isTryptic'.}

\item{nameFT}{A character string defining how full-tryptic peptides are
annotated in the \code{infoTryptic} colunm in \code{annotPP}. Only peptides
annotated accordingly will remain in the data if \code{filterTryptic} is set
to 'TRUE'.
Default is 'Specific'.}
}
\value{
Returns a list of matrices containing half-tryptic LiP quantities of
interest (e.g. peptide, modified peptide, precursor), the best matching
full-tryptic TrP quantities of interest (e.g. peptide, modified peptide,
precursor) and corresponding TrP protein quantities. The row names in all
matrices correspond to the half-tryptic LiP quantity ID.
}
\description{
Function for preprocessing data for 'HTonly' mode prior to
running the RUV models. Half-tryptic (HT) peptides in the LiP data will be
matched to the best correlating full-tryptic (FT) TrP peptide originating
from the same protein(s). If the peptide matches to multiple proteins, TrP
peptides from all of these proteins can be matched. All full-tryptic LiP
peptides are subsequently removed and TrP peptides and protein qunatity
names will be altered to match the names of the corresponding half-tryptic
peptides.
}
