% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterSpectroOut.R
\name{filterForRunHT}
\alias{filterForRunHT}
\title{Filtering data for running runHT mode of package}
\usage{
filterForRunHT(SpectroList, annotPP, namePepQuant, nameProtQuant)
}
\arguments{
\item{SpectroList}{A list of matrices, containing peptides/proteins
quantities. Rows represent features and columns refer to the samples. Can
simply be output from \code{ExtractDataFromSpectro}.}

\item{annotPP}{A data.frame with peptide and protein annotation. Rows are
features and must match to the row.names of SpectroList/QuantityMatrix.
Must include all columns needed for filtering of data.}

\item{namePepQuant}{A character string giving column of \code{annotPP} which
matches to the row.names of the peptide quantities.
Set to 'Peptide' by default.}

\item{nameProtQuant}{A character string giving column of \code{annotPP} which
matches to the row.names of the protein quantities.
Set to 'Protein' by default.}
}
\value{
A list containing LiPPep and TrPProt quantities
}
\description{
Function for preprocessing data before fitting models.
Filtering can be applied based on different features such as digest type,
proteotypic or number of missed cleavages.
}
