% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterSpectroOut.R
\name{filterForRunHTonly}
\alias{filterForRunHTonly}
\title{Filtering data for running runHTonly mode of package}
\usage{
filterForRunHTonly(
  SpectroList,
  annotPP,
  annotS,
  filterNA,
  maxNAperCondition,
  infoCondition,
  namePepQuant,
  nameProtQuant,
  infoTryptic,
  nameFT
)
}
\arguments{
\item{SpectroList}{A list of matrices, containing peptides/proteins
quantities. Rows represent features and columns refer to the samples. Can
simply be output from \code{ExtractDataFromSpectro}.}

\item{annotPP}{A data.frame with peptide and protein annotation. Rows are
features and must match to the row.names of SpectroList/QuantityMatrix.
Must include all columns needed for filtering of data.}

\item{annotS}{A data.frame object containing sample annotation. Rows are
samples and must match to columns of \code{SpectroList}. If \code{filterNA}
is set to 'TRUE', it needs to contain column about different
conditions/groups.}

\item{filterNA}{A boolean value defining peptides should be filtered based on
number of NAs. Default is 'TRUE'.}

\item{maxNAperCondition}{A numeric value, defining maximal number of NAs
in an individual feature per condition. Default is '0'.}

\item{infoCondition}{A character string or numeric giving the column name
or column number of \code{annotS} in which condition is provided. Default is
'Condition'. If NA filtering should not be applied on condition level provide
only one value in \code{infoCondition} in \code{annotS}.}

\item{namePepQuant}{A character string giving column of \code{annotPP} which
matches to the row.names of the peptide quantities.
Set to 'Peptide' by default.}

\item{nameProtQuant}{A character string giving column of \code{annotPP} which
matches to the row.names of the protein quantities.
Set to 'Protein' by default.}

\item{infoTryptic}{A character string or numeric giving the column name or
column number of \code{annotPP} in which digest type can be found
Default is 'isTryptic'.}

\item{nameFT}{A character string defining which annotation in
\code{infoTryptic} should remain in the data. Default is 'Specific'.}
}
\value{
A list containing LiPPep and TrpProt quantities
}
\description{
Function for preprocessing data before fitting models.
Filtering can be applied based on different features such as digest type,
proteotypic or number of missed cleavages.
}
