% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterSpectroOut.R
\name{filterNAsFromList}
\alias{filterNAsFromList}
\title{Filtering quantities based on number of NAs}
\usage{
filterNAsFromList(quantityList, filterNA, annotS, infoCondition, maxNAs)
}
\arguments{
\item{quantityList}{A list of matrices, containing quantities of interest
(e.g. peptide, modified peptide, precursor) and protein abundances. Rows
represent features and columns samples. Sample names between the different
matrices contained in the list should match. Not matching samples will be
removed.
Output from \code{extractMSData} or \code{extractSpectroData} is in
the correct format (except of potentially not matching sample names, then
these need to be adjusted manually.)
\code{quantityList} may include the following matrices
\itemize{
  \item 'LiPPep': LiP peptide quantities (or alternatively the modified
  peptide/ precursor/other quantities, dependent on \code{quantName} and
  \code{quantValue})
  \item 'TrPPep': TrP peptide quantities (or alternatively the modified
  peptide/precursor/other quantities, dependent on \code{quantName} and
  \code{quantValue})
  \item 'TrPProt': TrP protein quantities
  \item 'LiPProt': LiP protein quantities
  }}

\item{filterNA}{A character variable defining type of NA filtering applied.
Can be set to c('all', 'byCondition' or 'none').
\itemize{
  \item 'all': Filtering of NAs will be applied on all samples, independent
  of their condition. \code{maxNA} defines the number of NAs allowed per
  quantity over all samples.
  \item 'byCondition': Filtering of NAs will be applied on samples based on
  their condition.  \code{maxNA} defines the number of NAs allowed per
  quantity over all samples of each individual condition. It is advised to
  only use 'byCondition', if the condition you are providing is a
  categorical variable and not continuous.
  \item 'none': No NA filtering will be applied. This my cause issues if
  running RUV and/or contrast models subsequently.
  }}

\item{annotS}{A data.frame object containing sample annotation. Rows are
samples and must match to columns of \code{quantityList}.
If \code{filterNA='byCondition'} it needs to contain column providing the
conditions/groups.}

\item{infoCondition}{A character string or numeric giving the column name
or column number of \code{annotS} in which condition is provided. Required,
if \code{filterNA = 'byCondition'}.
Default is 'Condition'.}

\item{maxNAs}{A numeric value, defining maximal number of NAs allowed.
Default is '0'.}
}
\value{
\code{quantityList} filtered based on number of NAs per quantity ID
}
\description{
Function for removing quantities containing more NAs than
defined by the user on a joined quantity level. This reflects the number
of samples that cannot be modeled in the RUV model later.
}
