% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingAndSorting.R
\name{getPepProtAnnotSpectro}
\alias{getPepProtAnnotSpectro}
\title{Creating a data.frame with peptide and protein annotation.}
\usage{
getPepProtAnnotSpectro(spectroOut, spectroOut2=NULL,
analysisLvl="Peptide")
}
\arguments{
\item{spectroOut}{MS report exported including sample and quantity information,
e.g. Spectronaut report.}

\item{spectroOut2}{An optional additional MS report including sample and
quantity information, e.g. Spectronaut report. If included theh peptide &
protein annotation file will also include quantities uniquely measured in
\code{reportOut}. Default is 'NULL', assuming only one MS report being
presented.}

\item{analysisLvl}{A character string defining the level on which the
peptide/protein quantities should be exported. Peptide' by default, can
alternatively be set to 'ModifiedPeptide' or Precursor'. Should be identical
to the \code{analysisLvl} defined in \code{extractSpectroData}.}
}
\value{
Returns a data.frame including all necessary annotation on peptides
and proteins
}
\description{
Create an annotation data.frame for the quantities from a
Spectronaut report/Spectronaut reports. The here created annotation file is
provides a structured overview about peptide and protein features and is
required in further downstream analysis.
Data must have been exported using the Spectronaut schema
'SpectroSchema_LiPAnalyzerOut', else, please use the function
\code{extractMSData}.
}
