% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingResults.R
\name{getPlottingFormat}
\alias{getPlottingFormat}
\title{Function for creating data.frame to create wood plot(s)}
\usage{
getPlottingFormat(sumDf, annotPP, coefCol, pvalCol, nameProtQuant,
startPosition, endPosition, pvalCutoff, deltaColorIsTryptic, isTryptic,
nameFT, nameHT)
}
\arguments{
\item{sumDf}{A data.frame containing coefficients and p-values of the
condition of interest were rows corresponding to the quantity of interest
(e.g. peptides).}

\item{annotPP}{A data.frame with peptides (/modified peptides/precursors) and
protein annotation. Rows are features and must contain the row names of
\code{sumDf}. Must include all columns needed for plotting the data:
\itemize{
  \item \code{nameProtQuant} protein IDs
  \item \code{startPosition} AA positions were peptides (/modified peptides
  /precursors) start in protein
  \item \code{endPosition} AA positions were peptides (/modified peptides
  /precursors) end in protein
  \item \code{isTryptic} optional column providing information if peptide is
  full-tryptic or half-tryptic, only needed if
  \code{deltaColorIsTryptic = 'TRUE'}
}}

\item{coefCol}{A character vector or numeric value defining column of
\code{sumDf} were coefficients of the quantity of interest (e.g. peptides)
are provided.}

\item{pvalCol}{A character vector or numeric value defining column of
\code{sumDf} were (adjusted) p-values  of the quantity of interest
(e.g. peptides) are provided.}

\item{nameProtQuant}{A character string giving column of \code{annotPP} which
protein names are provided. If a peptides matches to the same protein several
times, the protein name should be provided each time, separated by ','. If a
peptide maps to the multiple proteins, these different proteins can be
provided by separating them with ';'.}

\item{startPosition}{A character string or numeric giving the column name or
column number in which start position from each peptide in its protein
sequence are provided in \code{annotPP}. If a peptides matches to the sam
protein several times, different start positions should be separated by ','.
If a peptide maps to the multiple proteins, the start positions should be
separated with ';'.}

\item{endPosition}{A character string or numeric giving the column name or
column number in which end position from each peptide in its protein
sequence are provided in \code{annotPP}. If a peptides matches to the sam
protein several times, different start positions should be separated by ','.
If a peptide maps to the multiple proteins, the start positions should be
separated with ';'.}

\item{pvalCutoff}{A numeric value, peptides with p-values below this cut-off
are considered significant.}

\item{deltaColorIsTryptic}{A boolean variable, if set to 'TRUE', full-tryptic
and half-tryptic peptides will be plotted in different colors, this requires
\code{isTryptic} column in \code{annotPP}.}

\item{isTryptic}{A character string or numeric giving the column name or
column number of \code{annotPP} in which digest type can be found}

\item{nameFT}{A character vector defining the name of full-tryptic peptides
provided in the \code{isTryptic} column in \code{annotPP}.}

\item{nameHT}{A character vector defining the name of half-tryptic peptides
provided in the \code{isTryptic} column in \code{annotPP}.}
}
\value{
Data.frame based on which woods plots are created
}
\description{
Function for creating data.frame to create wood plot(s)
}
