% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingAndSorting.R
\name{joinPG}
\alias{joinPG}
\title{Combining protein names of identical peptides}
\usage{
joinPG(annotPP, iCol, protCol = "Protein")
}
\arguments{
\item{annotPP}{A data.frame with peptide and protein annotation.
Rows are feature and must include a column protein name(s)}

\item{iCol}{A character string or numeric giving the column name or
column number in which features are provided in the Spectronaut report
based on the chosen \code{analysisLvl}.}

\item{protCol}{A character string or numeric giving the column name or
column number in which protein name(s)  are provide. Set to 'Protein' by
default.}
}
\value{
A data.frame including all necessary annotation on peptides and
proteins with joined protein names.
}
\description{
Joining protein names together if a peptide was matched to
different PG groups in different Spectronaut reports.
}
