% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingResults.R
\name{makePCA}
\alias{makePCA}
\title{Creating PCA plot}
\usage{
makePCA(dataMat, annotS, infoCondition="Condition",
typeCondition="categorical", runProt=FALSE, annotPP=NULL,
nameProtQuant="Protein", Xaxis="PC1", Yaxis="PC2", nPCs=5)
}
\arguments{
\item{dataMat}{A matrix or data.frame providing data for estimating the PCA.
Rows are features and columns are samples. Can be a LiP or TrP peptide or
protein matrix (e.g. one element of the \code{preprocessQuantityMatrix}
output) or a matrix of residuals from the RUV model estimated with
\code{runModel}\\code{analyzeLiPPepData}\\code{analyzeTrPPepData}\
\code{analyzeProtData}.}

\item{annotS}{A data.frame containing sample annotation. Rows are samples and
must match to columns of the matrices in \code{quantityList}. Must include
column matching to \code{formulaContrast}.}

\item{infoCondition}{A character string providing column name of
\code{annotS} were condition of interest is provided. PCA will be colored
according to condition.
Default is 'Condition'.}

\item{typeCondition}{A character string providing information if
\code{inofCondition} is a 'categorical' or 'continuous' variable.
Default is 'categorical'.}

\item{runProt}{A boolean variable, if set to 'TRUE', function will remove
duplicated rows in \code{dataMat} and match row names of \code{dataMat} to
\code{nameProtQuant} in \code{annotPP}.
Default is 'FALSE'.}

\item{annotPP}{A data.frame with peptides (/modified peptides/precursors) and
protein annotation. Rows are features and row.names of \code{dataMat} must be
found here. \code{annotPP} must include a column named \code{nameProtQuant}
providing protein (group) names. The output from \code{getPepProtAnnot} can
be given here.}

\item{nameProtQuant}{A character string giving column of \code{annotPP} were
protein names are provided.
Default is 'Protein'.}

\item{Xaxis}{Numeric variable giving which PC should be visualized on the
X-axis. Cannot be be higher than the number of PCs (\code{nPCs}) being
estimated.
Default is '1'.}

\item{Yaxis}{Numeric variable giving which PC should be visualized on the
Y-axis. Cannot be be higher than the number of PCs (\code{nPCs}) being
estimated.
Default is '2'.}

\item{nPCs}{Numierc variable defining number of PCs to be estimated in PCA.
Default is '5'.}
}
\value{
PCA plot colored by condition
}
\description{
Estimating and plotting PCA from LiP data, TrP data or RUV
residuals.
}
