% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingResults.R
\name{makeVolcanoPlot}
\alias{makeVolcanoPlot}
\title{Creating volcano plot}
\usage{
makeVolcanoPlot(sumDf, coefCol="Coefficient", pvalCol="Padj",
pvalCutoff=0.05,xlim=NULL, ylim=NULL)
}
\arguments{
\item{sumDf}{A data.frame containing coefficients and p-values of the
condition of interest were rows corresponding to the quantity of interest
(e.g. peptides). Can be the output of \code{summarizeModelResults}.}

\item{coefCol}{A character vector or numeric value defining column of
\code{sumDf} were coefficients of the quantity of interest (e.g. peptides)
are provided.
Default is 'Coefficient'.}

\item{pvalCol}{A character vector or numeric value defining column of
\code{sumDf} were (adjusted) p-values  of the quantity of interest
(e.g. peptides) are provided.
Default is 'Padj'.}

\item{pvalCutoff}{A numeric value, peptides with p-values below this cut-off
are considered significant.
Default is '0.05'.}

\item{xlim}{A numeric vector of the length two defining the limits of the
x-axis of the plot. If set to NULL, limits of x-axis are chosen based on
coefficients provided in \code{sumDf}.}

\item{ylim}{A numeric vector of the length two defining the limits of the
y-axis of the plot. If set to NULL, the lower limit starts at 1 and the upper
limit is defined to the smallest provided p-value in \code{pvalCol}.}
}
\value{
Returns a volcano plot (ggplot) of coefficients and matching p-values
}
\description{
Creating volcano plot with model coefficients & corresponding
p-values
}
