% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingResults.R
\name{makeWoodsPlotSingleProtein}
\alias{makeWoodsPlotSingleProtein}
\title{Creating wood plots over a single protein}
\usage{
makeWoodsPlotSingleProtein(sumDf, annotPP, coefCol="Coefficient",
pvalCol="Padj", infoProtein="Protein", infoStart="startPosition",
infoEnd="endPosition", pvalCutoff=0.05, deltaColorIsTryptic=FALSE,
infoTryptic="isTryptic", nameFT=c("Specific"),
nameHT=c("Specific-C", "Specific-N"), protName=NULL,
showPv=FALSE)
}
\arguments{
\item{sumDf}{A data.frame containing coefficients and p-values of
peptides, with rows corresponding to the peptides. Can be the output of
\code{summarizeModelResults}}

\item{annotPP}{A data.frame with peptide and protein annotation. Rows are
peptides and must match to the row.names of \code{sumDf}. Can be output of
\code{getPepProtAnnot}.Must include all columns needed for plotting the data
(Protein name, start and stop position if each peptide, if
\code{deltaColorIsTryptic} is set to TRUE, information if each peptide is
full or half tryptic).}

\item{coefCol}{A character vector or numeric value defining column of
\code{sumDf} in which coefficients of the peptides are provided. Set to
'Coefficient' per default.}

\item{pvalCol}{A character vector or numeric value defining column of
\code{sumDf} in which p-values of the peptides are provided. Set to
'Padj' per default.}

\item{infoProtein}{A character vector or numeric value defining column of
\code{annotPP} in which the protein name of the peptides are provided. Set to
'Protein' per default. If a peptide is mapped and should be plotted to
multiple proteins, these should be seperated within the column by ';'.}

\item{infoStart}{A character vector or numeric value defining column of
\code{annotPP} in which the start position the peptides are provided. Set to
'startPosition' per default. If a peptide is mapped and should be plotted to
multiple proteins, these should be seperated within the column by ';'. If a
peptide matches to multiple positions within the same protein, seperate by
','.}

\item{infoEnd}{A character vector or numeric value defining column of
\code{annotPP} in which the end position the peptides are provided. Set to
'startPosition' per default. If a peptide is mapped and should be plotted to
multiple proteins, these should be seperated within the column by ';'. If a
peptide matches to multiple positions within the same protein, seperate by
','.}

\item{pvalCutoff}{A numeric value, peptides with p-values below this cut-off
are considered to be significant. Default is set to '0.05'.}

\item{deltaColorIsTryptic}{A boolean variable, if set to 'TRUE', full and
half tryptic peptides will be plotted in different colors. Default is
'FALSE'.}

\item{infoTryptic}{A character string or numeric giving the column name or
column number of \code{annotPP} in which digest type can be found
Default is 'isTryptic'.}

\item{nameFT}{A character vector defining the name of fully tryptic peptides
provided in the \code{isTryptic} column in \code{annotPP}. Default is set to
c("Specific").}

\item{nameHT}{A character vector defining the name of half tryptic peptides
provided in the \code{isTryptic} column in \code{annotPP}. Default is set to
c("Specific-C", "Specific-N").}

\item{protName}{A character vector giving name of protein to be plotted.}

\item{showPv}{A boolean value defining if p-values should be displayed in
plots. Default is set to 'FALSE'.}
}
\description{
Creating wood plots over a single protein
}
