% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingResults.R
\name{plottingProtein}
\alias{plottingProtein}
\title{Function for plotting woods plot for single protein}
\usage{
plottingProtein(plotData, deltaColorIsTryptic, xlim=NULL, ylim=NULL,
showPv, export)
}
\arguments{
\item{plotData}{A data.frame to create the woods plot for one protein from,
created using the function \code{getPlottingFormat}.}

\item{deltaColorIsTryptic}{A boolean variable, if set to 'TRUE', full-tryptic
and half-tryptic peptides will be plotted in different colors, this requires
\code{isTryptic} column in \code{annotPP}.}

\item{xlim}{A numeric vector of the length two defining the limits of the
x-axis of the plot. If set to NULL, limits of x-axis are chosen based on
sequence positions of the peptides.}

\item{ylim}{A numeric vector of the length two defining the limits of the
y-axis of the plot. If set to NULL, limits of x-axis are chosen based on
the coefficients provided in \code{sumDf}.}

\item{showPv}{A boolean value, if set to 'TRUE' p-values as defined in
\code{pvalCol} should be displayed in plots.}

\item{export}{A boolean value defining if plots should directly be exported.
If set to 'TRUE', plots will be written into a .pdf file, displaying three
plots per sheet. Additionally, a list with the plots will still be returned
in R.}
}
\value{
Woods plot for peptides over a protein
}
\description{
Function for plotting woods plot for single protein
}
