% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModels.R
\name{runContrast}
\alias{runContrast}
\title{Running contrast models for every quantity provided}
\usage{
runContrast(modelMat)
}
\arguments{
\item{modelMat}{A list of model matrices to perform contrast modeling using
ordinary least square regression.}
}
\value{
Returns a list of all contrast models
}
\description{
Function to run contrast on a list of model matrices, one element in
the list should represent one peptide or protein.
}
