% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModels.R
\name{RunModel}
\alias{RunModel}
\title{Fitting models to MS data}
\usage{
RunModel(spectroList, annotS = NULL, formulaBVLS = "Y~XPep+XProt", formulaOLS = "Y~Condition",
lowBVLS = c(-1e9, 0, 0), upBVLS = c(Inf, Inf, Inf), addBVLSbounds = F,
returnBVLSmodels = F, returnOLSmodels = F, LiPonly = F)
}
\arguments{
\item{spectroList}{a list of matrices, containing peptide/protein quantities.
Rows represent features and columns refer to the samples.
Names of the list items should be set to "LiPPep", "TrpPep" and "TrpProt"
(if the LiP only version is run, names should refer to "LiPPep" and "LiPProt")}

\item{annotS}{a data.frame object containing sample annotation.
Rows are samples and must match to columns of \code{spectroList}.}

\item{formulaBVLS}{a character string or formula defining the bounded variable least square models.
Use Y for defining the quantity matrix with values to predict. ´
If additional quantity matrices should be used as variables in the models,
please refer to these as XPep and/or XProt.
All other variables in the formula should refer to columns in \code{annotS}.
If BVLS should not be run, set to 'NULL'.
Default is defined as 'Y~XPep+XProt'.}

\item{formulaOLS}{a character string or formula defining the ordinary variable least square models.
Use Y for defining the quantity matrix with values to predict. ´
If additional quantity matrices should be used as variables in the models,
please refer to these as XPep and/or XProt.
All other variables in the formula should refer to columns in \code{annotS}.
If OLS should not be run, set to 'NULL'.
Default is defined as 'Y~Condition'.}

\item{lowBVLS}{a numeric vector defining lower boundaries of the coefficients of the BVLS models.
Elements refer to definition of \code{formulaBVLS}.
Default is defined as 'c(-Inf, 0, 0)'}

\item{upBVLS}{a numeric vector defining upper boundaries of the coefficients of the BVLS models.
Elements refer to definition of \code{formulaBVLS}.
Default is defined as 'c(Inf, Inf, Inf)'}

\item{addBVLSbounds}{a boolean value, if TRUE as many bounds as additionally needed
in each BVLS model are added to \code{lowBVLS} and \code{upBVLS}.
Added boundaries are automatically set to -Inf for \code{lowBVLS} and Inf for \code{upBVLS}}

\item{returnBVLSmodels}{a boolean value, set to 'TRUE' for
additionally return all BVLS models}

\item{returnOLSmodels}{a boolean value, set to 'TRUE' for
additionally return all OLS models}

\item{LiPonly}{a boolean value, set to 'TRUE' if you are running the
LiPonly version o the package and not providing trypsin-only data}
}
\description{
Function to build linear regression models for fitting MS data

add later
}
