% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModels.R
\name{runRUV}
\alias{runRUV}
\title{Running RUV models for every quantity provided}
\usage{
runRUV(formula, modelMat, lowRUV, upRUV, addRUVbounds)
}
\arguments{
\item{formula}{A formula used to create \code{modelMat} for RUV models.}

\item{modelMat}{A list of model matrices to perform RUV using bounded
variable least square regression.}

\item{lowRUV}{A numeric vector defining lower boundaries of the coefficients
of the RUV models. Elements refer to definition of \code{formulaRUV}.
Default is 'c(-Inf, 0, 0)'.}

\item{upRUV}{A numeric vector defining upper boundaries of the coefficients
of the RUV models. Elements refer to definition of \code{formulaRUV}.
Default is 'c(Inf, Inf, Inf)'.}

\item{addRUVbounds}{A boolean value, if set to 'TRUE' as many bounds as
additionally needed based on \code{formulaRUV} in each RUV model are added to
\code{lowRUV} and \code{upRUV}. Added boundaries are automatically set to
\code{lowRUV = -Inf} and \code{upRUV = Inf}. Important to set to 'TRUE', if
you have categories with multiple levels in the RUV model and did not adjust
the RUV boundaries based in the number of levels. This might be the case if
you have more than two batches you aim to account for in the RUV model.
Default is 'FALSE'.}
}
\value{
Returns a list of all RUV models
}
\description{
Function to run RUV on a list of model matrices, one element in
the list should represent one peptide or protein.
}
