% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingResults.R
\name{separatePeptides}
\alias{separatePeptides}
\title{Function for duplicating peptides if they occur multiple times in the
same protein}
\usage{
separatePeptides(plotData, deltaColorIsTryptic)
}
\arguments{
\item{plotData}{Data.frame for plotting as created within
\code{getPlottingFormat}.}

\item{deltaColorIsTryptic}{A boolean variable, if set to 'TRUE', full-tryptic
and half-tryptic peptides will be plotted in different colors, this requires
\code{isTryptic} column in \code{annotPP}.}
}
\value{
\code{plotData} data.frame
}
\description{
Duplicatig rows of \code{plotData} based on how many start
positions the respective quantID as in the same protein. Adjusting rest of
\code{plotData} format accordingly.
}
