% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingResults.R
\name{separateProteins}
\alias{separateProteins}
\title{Function for duplicating peptides if they occur in multiple proteins}
\usage{
separateProteins(plotData, deltaColorIsTryptic)
}
\arguments{
\item{plotData}{Data.frame for plotting as created within
\code{getPlottingFormat}.}

\item{deltaColorIsTryptic}{A boolean variable, if set to 'TRUE', full-tryptic
and half-tryptic peptides will be plotted in different colors, this requires
\code{isTryptic} column in \code{annotPP}.}
}
\value{
\code{plotData} data.frame
}
\description{
Duplicatig rows of \code{plotData} based on how often the
respective quantID occurs in different proteins. Adjusting rest of
\code{plotData} format accordingly.
}
