% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelResults.R
\name{summarizeModelResults}
\alias{summarizeModelResults}
\title{Extracting model results and performing FDR correction on p-values.}
\usage{
summarizeModelResults(resModel, evalCovariable="Condition_Contrast",
correctPval="all", pAdjust="fdr", annotPP=NULL, nameIDQuant="quantID",
nameProtQuant="Protein")
}
\arguments{
\item{resModel}{A list object, output from \code{runModel}\
\code{AnalyzeLiPPepData}\\code{AnalyzeTrPPepData}\\code{AnalyzeTrPProtData}.
Can only be run if the contrast model was run in the function.}

\item{evalCovariable}{A character string providing name of columns in which
coefficients and p-values of the variable of interest can be found.
Default is 'Condition_Contrast'.}

\item{correctPval}{A character string defining approach to adjust p-values.
\itemize{
  \item 'all': All p-values are FDR adjusted together.
  \item 'protein-wise': FDR adjustment is performed in smaller groups,
  meaning over all peptides matching to the same protein
  }}

\item{pAdjust}{A character string giving the correction method to use when
performing FDR Check '?p.adjust' for more information.
Default is 'fdr' (Benjamini & Hochberg (1995)).}

\item{annotPP}{A data.frame with peptides (/modified peptides/precursors) and
protein annotation that must be provided if \code{correctPval} = 'all'. Rows
are features. Data.frame must contain columns providing the
\code{nameIDQuant} and matching protein names in \code{nameProtQuant}.
The output from \code{getPepProtAnnot} can be given here.
Default is 'NULL'.}

\item{nameIDQuant}{A character string giving column name of \code{annotPP}
where quantity IDs (e.g. peptide identifies/sequences) used in row names of
\code{resModel}.}

\item{nameProtQuant}{A character string giving column of \code{annotPP} which
protein names/protein groups are provided. If \code{correctPval} is set to
'protein-wise', peptides with the same \code{nameProtQuant} value will be
corrected together.
Default is 'Protein'.}
}
\value{
A data.frame were rows are features and columns are coefficient,
p-value and adjusted p-value of the condition of interest.
}
\description{
Function for extracting p-values and coefficients of interest
from the model output (\code{runModel}/\code{analyzeLiPPepData}/
\code{analyzeTrPPepData}/\code{analyzeTrPProtData}) output. The function
additionally performs FDR correction, either over all proteins or
protein-wise (meaning over all peptides matching to the same protein).
}
