% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModels.R
\name{analyzeLiPPepData}
\alias{analyzeLiPPepData}
\title{Fitting model to quantify  accessibility changes in LiP quantities}
\usage{
analyzeLiPPepData(quantityList, annotS, infoCondition="Condition",
formulaRUV="Y~XPep+XProt", formulaContrast, lowRUV=c(-1e9,0,0),
upRUV=c(Inf, Inf, Inf), addRUVbounds=FALSE, mode="default")
}
\arguments{
\item{quantityList}{A list of preprocessed matrices, containing quantities of
interest(e.g. peptide, modified peptide, precursor) and protein abundances.
Rows represent features and columns samples and should match between the
different matrices contained in the list.
Output from \code{preprocessQuantityMatrix} is in the correct format.
The matrices of \code{quantityList} may have the following names
\itemize{
  \item c('LiPPep', 'TrPPep', 'TrPProt') for
  \code{mode = c("default", "HTonly")}
  \item c('LiPPep', 'TrPProt') for \code{mode = "FTHTjoin"}
  \item c('LiPPep', 'LiPProt') for \code{mode = "LiPonly"}
  }}

\item{annotS}{A data.frame containing sample annotation. Rows are samples and
must match to columns of the matrices in \code{quantityList}. Must include
columns of any further variables used in \code{formulaRUV} and
\code{formulaContrast}, including \code{infoCondition}.}

\item{infoCondition}{A character string providing column name of
\code{annotS} were condition to fit in the contrast model is provided.
Default is 'Condition'.}

\item{formulaRUV}{A character string or formula defining the RUV models.
Default is 'Y~XPep+XProt'.}

\item{formulaContrast}{A character string or formula defining the contrast
models.
Default is 'NULL', causing the function to set \code{formulaContrast} to
'Y~\code{infoCondition}.}

\item{lowRUV}{A numeric vector defining lower boundaries of the coefficients
of the RUV models. Elements refer to definition of \code{formulaRUV}.
Default is 'c(-Inf, 0, 0)'.}

\item{upRUV}{A numeric vector defining upper boundaries of the coefficients
of the RUV models. Elements refer to definition of \code{formulaRUV}.
Default is 'c(Inf, Inf, Inf)'.}

\item{addRUVbounds}{A boolean value, if set to 'TRUE' as many bounds as
additionally needed based on \code{formulaRUV} in each RUV model are added to
\code{lowRUV} and \code{upRUV}. Added boundaries are automatically set to
\code{lowRUV = -Inf} and \code{upRUV = Inf}. Important to set to 'TRUE', if
you have categories with multiple levels in the RUV model and did not adjust
the RUV boundaries based in the number of levels. This might be the case if
you have more than two batches you aim to account for in the RUV model.
Default is 'FALSE'.}

\item{mode}{A character variable defining mode in which function is run. Can
be set to c('default', 'HTonly', 'FTHTjoin' or 'LiPonly').
\itemize{
  \item 'default': Correcting full-tryptic LiP quantities (e.g. peptides) for
  TrP peptide and protein quantities.
  \item 'HTonly': Correcting half-tryptic LiP quantities (e.g. peptides) for
  best matching TrP peptide and the corresponding protein quantities. Please
  run \code{preprocessQuantityMatrix} with \code{mode = 'HTonly'} to
  preprocess the \code{quantityList} prior to running the models.
  \item 'FTHTjoin': Correcting full-tryptic and half-tryptic LiP quantities
  (e.g. peptides) for TrP protein quantities. Please run
  \code{preprocessQuantityMatrix} with \code{mode = 'FTHTjoin'} to
  preprocess the \code{quantityList} prior to running the models.
  \item 'LiPonly': Correcting LiP quantities (e.g. peptides) for LiP protein
  quantities.
  }
When aiming to run different corrections (e.g. correct LiP peptide for TrP
peptide quantities) please use \code{runModel} instead. This function allows
for more concrete settings.}
}
\value{
If running RUV and contrast model a list containing the following
matrices is returned:
\itemize{
  \item Matrix containing coefficients estimated in both models.
  \item P-values estimated for coefficients of contrast model.
  \item Matrix with the residuals resulting from the RUV model, rows are
  features and columns are samples.
}
If only a contrast model is run a list containing the following matrices is
returned:
\itemize{
  \item Matrix containing coefficients estimated in the contrast model.
  \item P-values estimated for coefficients of contrast model.
}
}
\description{
Function to build linear regression models for fitting MS data
and retrieving structural variation between different conditions.
}
