% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModels.R
\name{AnalyzeTrpPepData}
\alias{AnalyzeTrpPepData}
\title{Fitting model for PK-independent changes in Trp peptides}
\usage{
AnalyzeTrpPepData(spectroList, annotS, infoCondition="Condition",
formulaBVLS="Y~XProt", formulaOLS=NULL, lowBVLS=c(-1e9, 0),
upBVLS=c(Inf, Inf), addBVLSbounds=FALSE)
}
\arguments{
\item{spectroList}{A list of matrices, containing peptide/protein quantities.
Rows represent features and columns refer to the samples. Names of the list
items should be set to "LiPPep", "TrpPep" and "TrpProt".}

\item{annotS}{A data.frame containing sample annotation. Rows are samples and
must match to columns of \code{spectroList}.}

\item{infoCondition}{A character string providing column name of
\code{annotS} in which condition is provided. Default is 'Condition'.}

\item{formulaBVLS}{A character string or formula defining the BVLS models.
Default is defined as 'Y~XProt'.}

\item{formulaOLS}{A character string or formula defining the OLS models.
If 'NULL' will be set to 'Y~\code{infoCondition}.}

\item{lowBVLS}{A numeric vector defining lower boundaries of the coefficients
of the BVLS models. Elements refer to definition of \code{formulaBVLS}.
Default is defined as 'c(-Inf, 0)'.}

\item{upBVLS}{A numeric vector defining upper boundaries of the coefficients
of the BVLS models. Elements refer to definition of \code{formulaBVLS}.
Default is defined as 'c(Inf, Inf)'.}

\item{addBVLSbounds}{A boolean value, if set to 'TRUE' as many bounds as
additionally needed in each BVLS model are added to \code{lowBVLS} and
\code{upBVLS}. Added boundaries are automatically set to -Inf for
\code{lowBVLS} and Inf for \code{upBVLS}. Important to set to 'TRUE', if you
are for example also running batch correction in the BVLS model.}
}
\description{
Function to build linear regression models for fitting MS data
and retrieving PK-independent peptide variation between different conditions.
}
