% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModels.R
\name{analyzeTrpPepData}
\alias{analyzeTrpPepData}
\title{Fitting model for PK-independent changes in Trp peptides}
\usage{
analyzeTrpPepData(spectroList, annotS, infoCondition="Condition",
formulaRUV="Y~XProt", formulaContrast=NULL, lowRUV=c(-1e9, 0),
upRUV=c(Inf, Inf), addRUVbounds=FALSE)
}
\arguments{
\item{spectroList}{A list of matrices, containing peptide/protein quantities.
Rows represent features and columns refer to the samples. Names of the list
items should be set to "LiPPep", "TrpPep" and "TrpProt".}

\item{annotS}{A data.frame containing sample annotation. Rows are samples and
must match to columns of \code{spectroList}.}

\item{infoCondition}{A character string providing column name of
\code{annotS} in which condition is provided. Default is 'Condition'.}

\item{formulaRUV}{A character string or formula defining the RUV models.
Default is defined as 'Y~XProt'.}

\item{formulaContrast}{A character string or formula defining the contrast
models. If 'NULL' will be set to 'Y~\code{infoCondition}.}

\item{lowRUV}{A numeric vector defining lower boundaries of the coefficients
of the RUV models. Elements refer to definition of \code{formulaRUV}.
Default is defined as 'c(-Inf, 0)'.}

\item{upRUV}{A numeric vector defining upper boundaries of the coefficients
of the RUV models. Elements refer to definition of \code{formulaRUV}.
Default is defined as 'c(Inf, Inf)'.}

\item{addRUVbounds}{A boolean value, if set to 'TRUE' as many bounds as
additionally needed in each RUV model are added to \code{lowRUV} and
\code{upRUV}. Added boundaries are automatically set to -Inf for
\code{lowRUV} and Inf for \code{upRUV}. Important to set to 'TRUE', if you
are for example also running batch correction in the RUV model.}
}
\description{
Function to build linear regression models for fitting MS data
and retrieving PK-independent peptide variation between different conditions.
}
