% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModels.R
\name{analyzeTrPPepData}
\alias{analyzeTrPPepData}
\title{Fitting model to quantify PK-independent changes in TrP peptides}
\usage{
analyzeTrPPepData(quantityList, annotS, infoCondition="Condition",
formulaRUV="Y~XProt", formulaContrast=NULL, lowRUV=c(-1e9, 0),
upRUV=c(Inf, Inf), addRUVbounds=FALSE)
}
\arguments{
\item{quantityList}{A list of preprocessed matrices, containing quantities of
interest(e.g. peptide, modified peptide, precursor) and protein abundances.
Rows represent features and columns samples and should match between the
different matrices contained in the list.
Output from \code{preprocessQuantityMatrix} is in the correct format.
The matrices of \code{quantityList} may have the following names
\itemize{
  \item c('LiPPep', 'TrPPep', 'TrPProt') - in this case, 'LiPPep' will not
  be taken into account.
  \item c('TrPPep', 'TrPProt)
  }}

\item{annotS}{A data.frame containing sample annotation. Rows are samples and
must match to columns of the matrices in \code{quantityList}. Must include
columns of any further variables used in \code{formulaRUV} and
\code{formulaContrast}, including \code{infoCondition}.}

\item{infoCondition}{A character string providing column name of
\code{annotS} were condition to fit in the contrast model is provided.
Default is 'Condition'.}

\item{formulaRUV}{A character string or formula defining the RUV models.
Default is defined as 'Y~XProt'.}

\item{formulaContrast}{A character string or formula defining the contrast
models.
Default is 'NULL', causing the function to set \code{formulaContrast} to
'Y~\code{infoCondition}.}

\item{lowRUV}{A numeric vector defining lower boundaries of the coefficients
of the RUV models. Elements refer to definition of \code{formulaRUV}.
Default is 'c(-Inf, 0)'.}

\item{upRUV}{A numeric vector defining upper boundaries of the coefficients
of the RUV models. Elements refer to definition of \code{formulaRUV}.
Default is 'c(Inf, Inf)'.}

\item{addRUVbounds}{A boolean value, if set to 'TRUE' as many bounds as
additionally needed based on \code{formulaRUV} in each RUV model are added to
\code{lowRUV} and \code{upRUV}. Added boundaries are automatically set to
\code{lowRUV = -Inf} and \code{upRUV = Inf}. Important to set to 'TRUE', if
you have categories with multiple levels in the RUV model and did not adjust
the RUV boundaries based in the number of levels. This might be the case if
you have more than two batches you aim to account for in the RUV model.
Default is 'FALSE'.}
}
\description{
Function to build linear regression models for fitting MS data
and retrieving PK-independent peptide variations between different
conditions.
}
