% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunningModels.R
\name{analyzeTrPProtData}
\alias{analyzeTrPProtData}
\title{Fitting model to quantify protein abundance changes}
\usage{
analyzeTrPProtData(quantityList, annotS, annotPP,
infoCondition="Condition", nameProtQuant="Protein", formulaRUV=NULL,
formulaContrast=NULL, lowRUV=NULL, upRUV=NULL, addRUVbounds=FALSE)
}
\arguments{
\item{quantityList}{A list of matrices, containing peptide/protein quantities.
Rows represent features and columns refer to the samples. Names of the list
items should be set to "LiPPep", "TrPPep" and "TrPProt" (if the LiP only
version is run, names should refer to "LiPPep" and "LiPProt").}

\item{annotS}{A data.frame containing sample annotation. Must contain all
columns included in the RUV and contrast models. Rows are samples and must
match to columns of the matrices in \code{quantityList}.}

\item{annotPP}{A data.frame with peptides (/modified peptides/precursors) and
protein annotation. Rows are features and must match to the row.names of the
quantityList. Must include a column named \code{nameProtQuant} providing
protein (group) names. Per default, the output from \code{getPepProtAnnot}
can be given here.}

\item{infoCondition}{A character string providing column name of
\code{annotS} were condition to fit in the contrast model is provided.
Default is 'Condition'.}

\item{nameProtQuant}{A character string giving column of \code{annotPP} were
protein names are provided..
Default is 'Protein'.}

\item{formulaRUV}{A character string or formula defining the RUV models.
Default is defined as 'NULL'.}

\item{formulaContrast}{A character string or formula defining the contrast
models.
Default is 'NULL', causing the function to set \code{formulaContrast} to
'Y~\code{infoCondition}.}

\item{lowRUV}{A numeric vector defining lower boundaries of the coefficients
of the RUV models. Elements refer to definition of \code{formulaRUV}.
Default is 'NULL'.}

\item{upRUV}{A numeric vector defining upper boundaries of the coefficients
of the RUV models. Elements refer to definition of \code{formulaRUV}.
Default is 'NULL'.}

\item{addRUVbounds}{A boolean value, if set to 'TRUE' as many bounds as
additionally needed based on \code{formulaRUV} in each RUV model are added to
\code{lowRUV} and \code{upRUV}. Added boundaries are automatically set to
\code{lowRUV = -Inf} and \code{upRUV = Inf}. Important to set to 'TRUE', if
you have categories with multiple levels in the RUV model and did not adjust
the RUV boundaries based in the number of levels. This might be the case if
you have more than two batches you aim to account for in the RUV model.
Default is 'FALSE'.}
}
\description{
Function to build linear regression models for fitting MS data
and retrieving protein abundance variation between different conditions.
}
