% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingAndSorting.R
\name{Convert2Matrix}
\alias{Convert2Matrix}
\title{Extracting a single peptude/protein matrix from Spectronaut report}
\usage{
Convert2Matrix(spectroOut, values, rows, cols)
}
\arguments{
\item{spectroOut}{Spectronaut report exported using the Spectronaut schema
'SpectroSchema_LiPAnalyzerOut'.}

\item{values}{A character string defining the column in Spectronaut report
where quantities to export are located/}

\item{rows}{A character string defining the column in Spectronaut report
where the names of the quantities to export are located. If there are
different quantities for the same \code{rows} the mean of these values is
used.}

\item{cols}{A character string defining the column in Spectronaut report
where the sample or file names are located.}
}
\value{
Retunr a matrix atrix with quantities, \code{rows} represent features
and \code{cols} refer to the samples.
}
\description{
This function exports quantities from the Spectronaut report
and writes them into a matrix. Rows are peptides/proteins and columns are
samples.
}
