% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingAndSorting.R
\name{convert2Matrix}
\alias{convert2Matrix}
\title{Extracting a single peptide/protein matrix from a MS report.}
\usage{
convert2Matrix(reportOut, quantValue, rows, cols)
}
\arguments{
\item{reportOut}{MS report as data.frame or matrix including sample and
quantity information, e.g. Spectronaut report.}

\item{quantValue}{A character string or numeric variable referring to the
column name or column number from the MS report in which quantities or
interest are provided. If there are multiple quantities referring to the same
\code{rows}, the mean of these will be estimated and returned.}

\item{rows}{A character string or numeric giving the column name or column
number in the  MS report where quantity IDs are provided.}

\item{cols}{A character string or numeric giving the column name or column
number in the MS report where sample names are provided.}
}
\value{
Returns a matrix with quantities, \code{rows} represent features
and \code{cols} refer to the samples.
}
\description{
Export quantities from the MS report, writing them into a
matrix were rows represent features and columns samples.
}
