% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelResults.R
\name{doProteinWiseFDR}
\alias{doProteinWiseFDR}
\title{Performing protein-wise FDR on peptide p-values}
\usage{
doProteinWiseFDR(vecPval, pAdjust, annotPP, infoFeature, infoProtein)
}
\arguments{
\item{vecPval}{A named numeric vector with p-values and corresponding
features}

\item{pAdjust}{A character string giving the correction method to use when
performing FDR Check ?p.adjust for more information. Set to 'fdr' (Benjamini
& Hochberg (1995)) by default.}

\item{annotPP}{A data.frame with peptide and protein annotation. Rows are
features and columns must include \code{infoFeature} and \code{infoProtein}.
Is only required if \code{correctPval} is set to 'protein-wise'.}

\item{infoFeature}{A character string providing column name of \code{annotPP}
where features which were fitted in the models are noted. Default is set to
'Peptide'.}

\item{infoProtein}{A character string providing column name of \code{annotPP}
where protein names/groups are noted. If \code{correctPval} is defines as
'protein-wise' peptides with the same \code{infoProtein} value will be
corrected together. Default is set to 'Protein'.}
}
\value{
vector of protein-wise adjusted P values
}
\description{
Performing protein-wise FDR on peptide p-values
}
