% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelResults.R
\name{doProteinWiseFDR}
\alias{doProteinWiseFDR}
\title{Performing protein-wise FDR on peptide p-values}
\usage{
doProteinWiseFDR(vecPval, pAdjust, annotPP, nameIDQuant, nameProtQuant)
}
\arguments{
\item{vecPval}{A named numeric vector with p-values, were names are the
corresponding features}

\item{pAdjust}{A character string giving the correction method to use when
performing FDR Check ?p.adjust for more information. Set to 'fdr' (Benjamini
& Hochberg (1995)) by default.}

\item{annotPP}{A data.frame with peptides (/modified peptides/precursors) and
protein annotation. Rows are features. Data.frame must contain
columns providing the \code{nameIDQuant} and matching protein names in
\code{nameProtQuant}. Per default, the output from \code{getPepProtAnnot}
can be given here.}

\item{nameIDQuant}{A character string giving column name of \code{annotPP}
where quantity IDs (e.g. peptide identifies/sequences) used in names of
\code{vecPval}.}

\item{nameProtQuant}{A character string giving column of \code{annotPP} which
protein names/protein groups are provided. If \code{correctPval} is set to
'protein-wise', peptides with the same \code{nameProtQuant} value will be
corrected together.
Default is 'Protein'.}
}
\value{
vector of protein-wise adjusted P values
}
\description{
Performing protein-wise FDR on peptide p-values
}
