% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingAndSorting.R
\name{ExtractDataFromSpectro}
\alias{ExtractDataFromSpectro}
\title{Creating  peptide and protein matrices from the Spectronaut report.}
\usage{
ExtractDataFromSpectro(spectroLiP, spectroTrp = NULL, analysisLvl = "Peptide",
sampleName = "R.FileName", valuePep = NULL,
valueProt = "PG.Quantity", LiPonly = F)
}
\arguments{
\item{spectroLiP}{Spectronaut report of LiP data.
Spectronaut report should have been exported using the Spectronaut schema SpectroSchema_LiPAnalyzerOut.}

\item{spectroTrp}{Spectronaut report of Trp data
Spectronaut report should have been exportet using the Spectronaut schema SpectroSchema_LiPAnalyzerOut.}

\item{analysisLvl}{a character defining the level on which the peptide/protein quantities should be exported.
'Peptide' by default, can alternatively be set to "ModifiedPeptide" or "Precursor".}

\item{sampleName}{a character defining the column of the Spectronaut report where samples names are provided.
'R.Filename' by default.}

\item{valuePep}{a character defining the column from which peptide/precursor quantities should be taken from.
NULL by default, automatically setting \code{valuePep} based on the \code{AnalysisLvl}.}

\item{valueProt}{a character defining the column from which protein quantities should be taken from.
'PG.Quantity' by default.}

\item{LiPonly}{a boolean value, set to 'TRUE' if you are running the
LiPonly version o the package and not providing trypsin-only data}
}
\value{
list of three matrices, containing the LiP peptide/precursor,
Trp peptide/precursor and Trp protein quantities.
Rows represent features and columns refer to the samples.
}
\description{
Function which exports peptide and protein quantities from a Spectronaut report and writes them into a matrix.
Data is exported from both the LiP and Trp Spectronaut reports per default.
Alternatively, function can be run in \code{LiPonly} mode, only requiring LiP data.
}
