% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingAndSorting.R
\name{ExtractDataFromSpectro}
\alias{ExtractDataFromSpectro}
\title{Creating peptide and protein matrices from the Spectronaut report.}
\usage{
ExtractDataFromSpectro(spectroLiP, spectroTrp=NULL,
analysisLvl="Peptide", sampleName="R.FileName", valuePep=NULL,
valueProt="PG.Quantity", LiPonly=FALSE)
}
\arguments{
\item{spectroLiP}{Spectronaut report of LiP data. Spectronaut report should
have been exported using the Spectronaut schema
'SpectroSchema_LiPAnalyzerOut'.}

\item{spectroTrp}{Spectronaut report of Trp data. Spectronaut report should
have been exported using the Spectronaut schema
'SpectroSchema_LiPAnalyzerOut'.}

\item{analysisLvl}{A character string defining the level on which the
peptide/protein quantities should be exported. Is set to 'Peptide' by d
efault, can alternatively be set to 'ModifiedPeptide' or 'Precursor'.}

\item{sampleName}{A character string defining the column of the Spectronaut
report where samples names are provided. Is set to 'R.Filename' by default.}

\item{valuePep}{A character string defining the column from which
peptide/precursor quantities should be taken from. Defined as 'NULL' by
default, resulting in the function automatically setting \code{valuePep}
based on the \code{AnalysisLvl} ('Peptide'='PEP.Quantity',
'ModifiedPeptide'='FG.Quantity', 'Precursor'='FG.Quantity').}

\item{valueProt}{A character string defining the column from which protein
quantities should be taken from. Set to 'PG.Quantity' by default.}

\item{LiPonly}{A boolean value, set to 'TRUE' if you want to run the LiPonly
version o the package and not providing trypsin-only data.Default is set to
'FALSE'.}
}
\value{
Per default a list of three matrices is returned, containing the LiP
peptide/precursor, Trp peptide/precursor and Trp protein quantities. If run
in \code{LiPonly} mode will return a list of two matrcies with the LiP
peptide/precursor and Trp protein quantities. Rows represent features and
columns refer to the samples.
}
\description{
This function writes peptide and protein quantities from a
Spectronaut report into a matrix.Data is exported from both the LiP and Trp
Spectronaut reports per default. Alternatively, function can be run in
\code{LiPonly} mode, only requiring LiP data.
}
