% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingAndSorting.R
\name{getEndPositionOfPep}
\alias{getEndPositionOfPep}
\title{Estimating end position of peptide sequences}
\usage{
getEndPositionOfPep(annotPP, startPosition="startPosition",
pepCol="Peptide")
}
\arguments{
\item{annotPP}{A data.frame with peptide and protein annotation were rows are
features. The data.frame must include a column providing the peptide
sequences and one providing the start position the peptide has in protein
sequence. Different start positions for the same peptide should be separated
by ',' for peptides mapping to the same protein multiple times and ';' for
peptides mapping to multiple proteins.}

\item{startPosition}{A character string or numeric giving the column name or
column number defining column in which start positions of peptides are
provided.
Default is 'startPosition'.}

\item{pepCol}{A character string or numeric giving the column name or
column number defining column in which peptide sequences are provide. Peptide
sequences have to be provided without modifications or charges.
Default is 'Peptide'.}
}
\value{
Returns a character vector with the end positions of all peptides.
Different start positions for the same peptide should be separated by ',' for
peptides mapping to the same protein multiple times and ';' for peptides
mapping to multiple proteins.
}
\description{
Getting end position of peptides sequences from the starting
positions provided. Function takes into account that peptides can map to
multiple regions in the proteome, providing multiple end positions if
multiple start positions are given.
}
