% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingAndSorting.R
\name{GetSampleAnnot}
\alias{GetSampleAnnot}
\title{Creating data.frame with sample annotation from Spectronaut report}
\usage{
GetSampleAnnot(spectroOut, sampleName = "R.FileName", sampleCondition = "R.Condition",
typeCondition = "factor", contrastCoding = "dummyCoding", baseLevel = NULL)
}
\arguments{
\item{spectroOut}{Spectronaut report.
Spectronaut report should have been exported using the Spectronaut schema SpectroSchema_LiPAnalyzerOut.}

\item{sampleName}{character giving the column in which sample names can be found in the Spectronaut report.
'R.FileName' by default.
Only used if \code{analysisLvl} is set to 'Precursor'.}

\item{sampleCondition}{character giving the column in which conditions which should be compared can be found in the Spectronaut report.
'R.Condition' by default.}

\item{typeCondition}{character providing information if condition is a factor or continuous variable
If condition is continuous has to be provided as numeric.
'factor' is default.}

\item{contrastCoding}{character providing information which type of contrast coding to use if condition is a factor
'dummyCoding' is default, alternatives 'sumCoding' or 'wecCoding' can be chosen
Can be set to NULL to not define contrast coding automatically}

\item{baseLevel}{character giving name of reference level if condition is a factor
and dummy or wec coding is used
If not provided will choose the first condition in Spectronaut report}
}
\value{
data.frame containing sample and condition annotation.
Rows are samples.
}
\description{
Creates an annotation file on samples including information on the conditions of samples.
This function only runs, if all the annotations were added to the Spectronaut report, else
please create the sample annotation file yourself.
}
