% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadingAndSorting.R
\name{getSampleAnnot}
\alias{getSampleAnnot}
\title{Creating data.frame with sample annotation from Spectronaut report}
\usage{
getSampleAnnot(spectroOut, sampleName="R.FileName",
sampleCondition="R.Condition", typeCondition="factor",
contrastCoding="dummyCoding", baseLevel=NULL)
}
\arguments{
\item{spectroOut}{Spectronaut report of MS data. Spectronaut report should
have been exported using the Spectronaut schema
'SpectroSchema_LiPAnalyzerOut'.}

\item{sampleName}{A character string  giving the column in which sample names
can be found in the Spectronaut report. Set to R.FileName' by default.}

\item{sampleCondition}{A character string giving the column in which
conditions can be found in the Spectronaut report. Set to 'R.Condition' by
default.}

\item{typeCondition}{A character string providing information if condition is
a factor or continuous variable If condition is continuous it has to be
provided as numeric in the Spectronaut report. Defined as ' 'factor' is
default, alternatively set to 'continuous'/}

\item{contrastCoding}{A character string providing information which type of
contrast coding to use if condition is a factor. Defined as 'dummyCoding' by
default. Alternatively, 'sumCoding' or 'wecCoding' can be chosen. Can be set
to NULL to not define contrast coding automatically in case you want to
defined it yourself.}

\item{baseLevel}{A character strong giving name of reference level if
\code{typeCondition} is a factor. If not provided the function will choose
the first condition in Spectronaut report as the reference level.}
}
\value{
Returns a data.frame containing sample and condition annotation.
Rows are samples.
}
\description{
Creates an annotation file on samples including information on
the conditions of samples. This function only runs, if all the annotations
were added to the Spectronaut report, else please create the sample
annotation file yourself.
}
