% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterSpectroOut.R
\name{preprocessQuantityMatrix}
\alias{preprocessQuantityMatrix}
\title{Preprocessing peptide and protein quantities}
\usage{
preprocessQuantityMatrix(SpectroList=NULL, QuantityMatrix=NULL,
annotPP=NULL, annotS=NULL, logT=TRUE, filterMinLogQuant=TRUE,
thresholdMinLogQuant=10, filterNA=TRUE, maxNAperCondition=0,
infoCondition="Condition", runHT = FALSE,  runHTonly = FALSE,
namePepQuant = "Peptide", nameProtQuant = "Protein", filterTryptic=TRUE,
infoTryptic="isTryptic", nameFT="Specific", filterProteotryptic=FALSE,
infoProteotypic="isProteotypic", nameProteotypic="True",
filterMissedCleaved=FALSE, maxMissedCleave=2,
infoMissedCleave="NMissedCleavages")
}
\arguments{
\item{SpectroList}{A list of matrices, containing peptides/proteins
quantities. Rows represent features and columns refer to the samples. Can
simply be output from \code{ExtractDataFromSpectro}.}

\item{QuantityMatrix}{A single matrix with peptide/protein quantities. Rows
represent features and columns refer to the samples.}

\item{annotPP}{A data.frame with peptide and protein annotation. Rows are
features and must match to the row.names of SpectroList/QuantityMatrix.
Must include all columns needed for filtering of data.}

\item{annotS}{A data.frame object containing sample annotation. Rows are
samples and must match to columns of \code{SpectroList}. If \code{filterNA}
is set to 'TRUE', it needs to contain column about different
conditions/groups.}

\item{logT}{A boolean value defining if data should be log-transformed.
Default is 'TRUE'.}

\item{filterMinLogQuant}{A boolean value defining small quantities should be
set to NA. Default is 'TRUE'.}

\item{thresholdMinLogQuant}{A numeric value, quantities below this threshold
will be set to NA if \code{filterMinLogQuant} is 'TRUE'. Default is set to
10. Setting it to 12 would be an optional stricter choice.}

\item{filterNA}{A boolean value defining peptides should be filtered based on
number of NAs. Default is 'TRUE'.}

\item{maxNAperCondition}{A numeric value, defining maximal number of NAs
in an individual feature per condition. Default is '0'.}

\item{infoCondition}{A character string or numeric giving the column name
or column number of \code{annotS} in which condition is provided. Default is
'Condition'. If NA filtering should not be applied on condition level provide
only one value in \code{infoCondition} in \code{annotS}.}

\item{runHT}{A boolean value defining if the function and further analysis
steps should be run including the half-tryptic LiP peptides. If set to true,
the TrpPep matrix will be removed, resulting in only the LiPPep and TrpProt
matrices remaining in the data and \code{filterTryptic} will automatically be
set to 'FALSE'. Need to provide \code{annotPP} to match protein quantities to
half-tryptic peptides in LiP data. \code{runHT} is set to 'FALSE' per
default.}

\item{runHTonly}{have to add description}

\item{namePepQuant}{A character string giving column of \code{annotPP} which
matches to the row.names of the peptide quantities.
Set to 'Peptide' by default.}

\item{nameProtQuant}{A character string giving column of \code{annotPP} which
matches to the row.names of the protein quantities.
Set to 'Protein' by default.}

\item{filterTryptic}{A boolean value defining if peptides should be filtered
based digest type. Default is 'TRUE', removing all not fully tryptic peptides
from the data matrices.}

\item{infoTryptic}{A character string or numeric giving the column name or
column number of \code{annotPP} in which digest type can be found
Default is 'isTryptic'.}

\item{nameFT}{A character string defining which annotation in
\code{infoTryptic} should remain in the data. Default is 'Specific'.}

\item{filterProteotryptic}{A boolean value, defining if peptides should be
filtered based on if they are proteotypic. Default is 'FALSE'.}

\item{infoProteotypic}{A character string or numeric giving the column name
or column number of \code{annotPP} in which annotation if peptides is
proteotypic is provided. Default is 'isProteotypic'.}

\item{nameProteotypic}{A character string defining which peptide in
\code{infoProteotypic} is proteotypic and will be filtered out if
\code{filterProteotryptic} is set to 'TRUE'. Default is 'True'.}

\item{filterMissedCleaved}{A boolean value defining if peptides should be
filtered based on number of missed cleavages. Default is 'FALSE'.}

\item{maxMissedCleave}{A numeric value, defining maximal number of missed
cleavages allowed per peptide if \code{filterMissedCleaved} is set to 'TRUE'.
Default is '2'.}

\item{infoMissedCleave}{A character string or numeric giving the column name
or column number of \code{annotPP} in which number of missed cleavage are
provided. Default is 'NMissedCleavages'.}
}
\value{
Returns a list of matrices containing preprocessed and filtered
peptide and protein quantities  OR a single matrix containing preprocessed
and filtered peptide or protein quantities. Rows represent features and
columns refer to the samples.
}
\description{
Function for preprocessing data before fitting models. Different
filtering can be applied based on various features such as digest type,
proteotypic status or number of missed cleavages.
}
