% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelResults.R
\name{SummarizeModelResults}
\alias{SummarizeModelResults}
\title{Extracting model results and performing FDR correction}
\usage{
SummarizeModelResults(resModel, evalCovariable = "Condition_OLS",
correctPval = "all", pAdjust = "fdr",
annotPP = NULL, infoFeature = "Peptide", infoProtein = "Protein")
}
\arguments{
\item{resModel}{a list object, output from runModel function.
\\code{AnalyzeLiPPepData}\\code{AnalyzeTrpPepData}\\code{AnalyzeTrpProtData}.}

\item{evalCovariable}{character string giving name of columns in which coefficients and
P values of the variable of interest can be found.
Default is set to 'Condition_OLS'.}

\item{correctPval}{a character string defining how p-values should be adjusted.
Per default all p-values are corrected together, and it is set to 'all'.
Alternatively, FDR can be performed protein-wise, please set the variable to
'protein-wise' for this.}

\item{pAdjust}{a character string giving the correction method to use when performing FDR
Check ?p.adjust for more information.
Set to 'fdr' (Benjamini & Hochberg (1995)) by default.}

\item{annotPP}{a data.frame with peptide and protein annotation.
Rows are features and columns must include \code{infoFeature} and \code{infoProtein}.
Is only required if \code{correctPval} is set to 'protein-wise'.}

\item{infoFeature}{a character string providing column name of \code{annotPP} where
features which were fitted in the models are noted.
Default is set to 'Peptide'.}

\item{infoProtein}{a character string providing column name of \code{annotPP} where
protein names/groups are noted. If \code{correctPval} is defines as 'protein-wise'
peptides with the same \code{infoProtein} value will be corrected together
Default is set to 'Protein'.}
}
\value{
returns a data.frame including all features of the model and
the corresponding coefficients, P values and adjusted P values.
}
\description{
Function for extracting p-values and coefficients from \code{RunModel} output.
Also performs FDR correction, either over all peptides/proteins or
over all peptides of each protein individually
}
